refactor: move cli into its own module (#7)
