+++ src/cli/mod.rs
@@ -1,0 +1,26 @@
+use clap::{Parser, Subcommand};
+
+mod roll;
+
+#[derive(Parser)]
+#[command(version, about, long_about = None)]
+struct Cli {
+    #[command(subcommand)]
+    command: Commands,
+}
+
+#[derive(Subcommand)]
+enum Commands {
+    /// Rolls some dice
+    Roll(roll::RollArgs),
+    /// Renders markdown files to HTML
+    Render,
+}
+
+pub fn run() {
+    let cli = Cli::parse();
+    match &cli.command {
+        Commands::Roll(args) => args.run(),
+        Commands::Render => todo!(),
+    }
+}
